* Initial implementation of the multi-arch virtual machine   - Allows to emulate register changes while disassembling or debugging   - Register Jump/Call/interrupt vector prediction   - Single interface for register name resolution   - add arch/x86/vm.c   - ATM is very basic interface * Initial integration of the x86 code analyzer with the vm * Make the 'ab' command accessible from the radare lua api   - code_block = Radare.Analyze.block(0x8049430)   print("Number of calls: "..code_block['n_calls']) * 'm'ove command shows help before checking file.write

+++ src/grava.c
@@ -431,7 +431,7 @@ void grava_program_graph(struct program_t *prg, struct mygrava_window *win)
 			grava_node_add_call(node, c0->addr);
 		}
 
-		node->baseaddr = (unsigned long)b0->addr;
+		node->baseaddr = b0->addr;
 		grava_node_set(node, "label", cmd+128);
 
 		/* disassemble body */
